requirements.txt
torch
torchvision
Pillow
Flask